refactor(server-db): not connecting directly to MongoDB Atlas on the server side
When I started writing complex logic in NestJS, I had to handle response requests, business logic,
database requests, data format transformations on the return...This is too coupled and reduces a lot
of code quality. I may use TRPC specifically for interacting with the database, similar to the DAO
layer.